some functions were renamed

+++ libraries/chain/controller.cpp
@@ -1409,12 +1409,12 @@ struct controller_impl {
          if( !validating ) {
             auto& rl = self.get_mutable_resource_limits_manager();
             rl.update_account_usage( trx_context.bill_to_accounts, block_timestamp_type(self.pending_block_time()).slot );
-            int64_t account_cpu_limit = 0;
+            uint32_t account_cpu_limit = 0;
             std::tie( std::ignore, account_cpu_limit, std::ignore, std::ignore ) = trx_context.max_bandwidth_billed_accounts_can_pay( true );
 
-            uint32_t limited_cpu_time_to_bill_us = static_cast<uint32_t>( std::min(
-                  std::min( static_cast<int64_t>(cpu_time_to_bill_us), account_cpu_limit ),
-                  trx_context.initial_objective_duration_limit.count() ) );
+            uint32_t limited_cpu_time_to_bill_us =  std::min(
+                  std::min( static_cast<uint32_t>(cpu_time_to_bill_us), account_cpu_limit ),
+                  static_cast<uint32_t>( trx_context.initial_objective_duration_limit.count() ) );
             EOS_ASSERT( !explicit_billed_cpu_time || (cpu_time_to_bill_us == limited_cpu_time_to_bill_us),
                         transaction_exception, "cpu to bill ${cpu} != limited ${limit}", ("cpu", cpu_time_to_bill_us)("limit", limited_cpu_time_to_bill_us) );
             cpu_time_to_bill_us = limited_cpu_time_to_bill_us;
